requirements.txt
torch==2.5.0
torchvision
einops
fire
huggingface-hub
safetensors
sentencepiece
transformers
tokenizers
protobuf
requests
invisible-watermark
pyyaml
python-box
ruff
PyYAML
loguru
ray
imageio[ffmpeg]
diffusers
accelerate
ftfy
flash_attn
easydict
moviepy==1.0.3
pillow==9.5.0
av==13.1.0
image-reward
git+https://github.com/openai/CLIP.git